+++ flake.nix
@@ -1,7 +1,8 @@
 {
   description = "A `flake-parts` module for Haskell development";
-  outputs = { self, ... }: {
-    flakeModule = ./nix/flake-module.nix;
+  inputs.nix-parsec.url = "github:kanwren/nix-parsec";
+  outputs = { self, nix-parsec, ... }: {
+    flakeModule = import ./nix/flake-module.nix { inherit nix-parsec; };
     templates.default = {
       description = "A simple flake.nix using haskell-flake";
       path = builtins.path { path = ./example; filter = path: _: baseNameOf path == "flake.nix"; };